test most features

+++ src/Pages/Home.tsx
@@ -11,15 +11,18 @@ import AccordionComponent from "../Components/UserInterface/Accordian";
 import Board from "../Components/UserInterface/Board";
 import AddEditTask from "../Components/UserInterface/AddEditTask";
 import { auth, db } from "../Firebase/firebase";
-import { doc, setDoc, serverTimestamp, collection,deleteDoc, getDocs, query, where, orderBy, updateDoc } from "firebase/firestore";
+import { doc, setDoc, serverTimestamp, collection, deleteDoc, getDocs, query, where, orderBy, updateDoc } from "firebase/firestore";
 import { toast } from 'react-toastify';
 import { useState, useEffect } from "react";
-import { taskCategories } from "../Utils/data";
+// import { taskCategories } from "../Utils/data";
 import { Select, Option, Typography, Input } from "@material-tailwind/react"
 import { MagnifyingGlassIcon } from "@heroicons/react/24/outline";
 import { startAt, endAt } from "firebase/firestore";
 import { useAuth } from "../Hooks/auth";
-
+import { ToastContainer } from 'react-toastify';
+import DatePicker from "react-datepicker";
+// import { Timestamp } from "firebase/firestore";
+const categories = ["Work", "Personal"];
 interface Task {
   id: string;
   taskStatus: string;
@@ -38,11 +41,28 @@ export default function Home() {
   const [tasks, setTasks] = useState<Task[]>([]);
   const [editTaskData, setEditTaskData] = useState({});
   const [payload, setPayload] = useState({
-    taskStatus:'',
-    taskTitle:''
+    taskCategory: '',
+    taskTitle: '',
+    startDate: null as Date | null,  // Ensure this is Date | null
+    endDate: null as Date | null     // Ensure this is Date | null
   });
-  const  user:any  = useAuth();
+  const user: any = useAuth();
+  const [startDate, setStartDate] = useState<Date | null>(null);
+  const [endDate, setEndDate] = useState<Date | null>(null);
 
+
+  const handleDateChange = (dates: [Date | null, Date | null]) => {
+    const [start, end] = dates;
+    setStartDate(start);
+    setEndDate(end);
+    if (end !== null) {
+      // console.log(end);
+      handleFilterChange("startDate", start);
+      handleFilterChange("endDate", end);
+    }
+
+
+  };
 
   useEffect(() => {
     fetchDocuments();
@@ -51,19 +71,32 @@ export default function Home() {
   const toggleModal = () => setIsModalOpen(!isModalOpen);
 
   const fetchDocuments = async () => {
+
+    // console.log(payload);
+
     try {
       const taskDataRef = collection(db, "taskdata");
 
       let filters: any[] = [];
       filters.push(where("uid", "==", user.uid));
-      if (payload.taskStatus) {
-        filters.push(where("taskStatus", "==", payload.taskStatus));
+      if (payload.taskCategory) {
+        filters.push(where("taskCategory", "==", payload.taskCategory));
       }
       if (payload.taskTitle) {
         filters.push(orderBy('taskTitle'));
         filters.push(startAt(payload.taskTitle));
         filters.push(endAt(payload.taskTitle + '\uf8ff'));
       }
+
+      // if (payload.startDate && payload.endDate) {
+      //   const startTimestamp = Timestamp.fromDate(new Date(payload.startDate)); // Convert to Date
+      //   const endTimestamp = Timestamp.fromDate(new Date(payload.endDate)); // Convert to Date
+
+      //   filters.push(orderBy("dueDate"));
+      //   filters.push(startAt(startTimestamp));
+      //   filters.push(endAt(endTimestamp));
+
+      // }
 
       let q;
       if (filters.length > 0) {
@@ -72,7 +105,7 @@ export default function Home() {
         q = taskDataRef;
       }
       const querySnapshot = await getDocs(q);
-      const data:any = querySnapshot.docs.map((doc) => {
+      const data: any = querySnapshot.docs.map((doc) => {
         const docData = doc.data();
         return {
           id: doc.id,
@@ -82,12 +115,34 @@ export default function Home() {
             month: 'long',
             day: 'numeric',
           }),
+          statusChangeTime: docData.statusChangeTime?new Date(docData.statusChangeTime.seconds * 1000).toLocaleDateString('en-US', {
+            year: 'numeric',
+            month: 'long',
+            day: 'numeric',
+          }):'',
+          timeStamp: new Date(docData.timeStamp.seconds * 1000).toLocaleDateString('en-US', {
+            year: 'numeric',
+            month: 'long',
+            day: 'numeric',
+          }),
         };
       });
 
-      setTasks(data);
+      if (payload.startDate && payload.endDate) {
+        const startDate = new Date(payload.startDate);
+        const endDate = new Date(payload.endDate);
+
+        const filteredTasks = data.filter((task: any) => {
+          const taskDueDate = new Date(task.date);
+          return taskDueDate >= startDate && taskDueDate <= endDate;
+        });
+
+        setTasks(filteredTasks);
+      } else {
+        setTasks(data);
+      }
       toast.success('Data Fetch Successfully');
-    } catch (error:any) {
+    } catch (error: any) {
       toast.error(error.message || "Error fetching tasks");
     }
   };
@@ -99,55 +154,64 @@ export default function Home() {
     }));
   };
 
+
+  const formatStatus = (status: string) => {
+    switch (status) {
+      case "TO_DO":
+        return "To-Do";
+      case "IN_PROGRESS":
+        return "In-Progress";
+      case "COMPLETED":
+        return "Completed";
+      default:
+        return status;
+    }
+  };
+
   //status Change
-  const handleStatusChange = async (newStatus: any, taskId: number) => {
+  const handleStatusChange = async (newStatus: any, taskId: number, prevTask: any) => {
     try {
+      const formattedPrevStatus = formatStatus(prevTask);
+      const formattedNewStatus = formatStatus(newStatus);
       const taskDocRef = doc(db, "taskdata", taskId.toString());
-      await updateDoc(taskDocRef, { taskStatus: newStatus });
+      await updateDoc(taskDocRef, {
+        taskStatus: newStatus,
+        statusChangeTrack: `You Changed status from ${formattedPrevStatus} to ${formattedNewStatus}`,
+        statusChangeTime: serverTimestamp()
+      });
       fetchDocuments();
     } catch (error) {
       console.error("Error updating task status:", error);
     }
   };
 
-  const handleEditTask =async (data:any)=>{
+  const handleEditTask = async (data: any) => {
     console.log(data);
     setEditTaskData(data)
     toggleModal();
   }
 
-  const handleResetEditTask=()=>{
+  const handleResetEditTask = () => {
     setEditTaskData({})
   }
 
   const handleDeleteTask = async (id: string) => {
-    try {
+    const isConfirmed = window.confirm('Are you sure you want to delete this task?');
+    if (isConfirmed) {
+      try {
         const taskDocRef = doc(db, 'taskdata', id);
         await deleteDoc(taskDocRef);
         fetchDocuments();
         toast.success('Task deleted successfully');
-    } catch (error:any) {
+      } catch (error: any) {
         toast.error('Error deleting task: ' + error.message);
         console.log(error);
+      }
+    } else {
+      toast.info('Task deletion canceled');
     }
-};
+  };
 
-  const data = [
-    {
-      label: "List",
-      value: "list",
-      icon: 'fa fa-list',
-      desc: <AccordionComponent tasks={tasks} updateStatus={handleStatusChange} onEdit={handleEditTask} ondelete={handleDeleteTask} />,
-    },
-    {
-      label: " Board",
-      value: "board",
-      icon: 'fa fa-list',
-      desc: <Board />
-    },
-  ];
-
- 
 
   //save task
   const handleSaveTask = async (data: any) => {
@@ -155,15 +219,15 @@ export default function Home() {
     const res: any = await auth.currentUser
     const taskId = data.id;
     try {
-      if(taskId){
+      if (taskId) {
         const taskDocRef = doc(db, 'taskdata', taskId);
         await updateDoc(taskDocRef, {
-            ...data,
-            uid: res.uid,
-            timeStamp: serverTimestamp(),
+          ...data,
+          uid: res.uid,
+          timeStamp: serverTimestamp(),
         });
         toast.success('Task updated successfully');
-      }else{
+      } else {
         const id = Date.now().toString()
         await setDoc(doc(db, 'taskdata', id), {
           ...data,
@@ -172,8 +236,7 @@ export default function Home() {
           timeStamp: serverTimestamp(),
         })
       }
-     
-      toggleModal();
+
       fetchDocuments();
       toast.success('Data Save Successfully');
       return true;
@@ -184,13 +247,30 @@ export default function Home() {
     }
   };
 
+  const data = [
+    {
+      label: "List",
+      value: "list",
+      icon: 'fa fa-list',
+      desc: <AccordionComponent tasks={tasks} updateStatus={handleStatusChange} onEdit={handleEditTask} ondelete={handleDeleteTask}
+        onSave={handleSaveTask} />,
+    },
+    {
+      label: " Board",
+      value: "board",
+      icon: 'fa fa-file',
+      desc: <Board tasks={tasks} onEdit={handleEditTask} ondelete={handleDeleteTask}
+        onUpdateTasks={handleStatusChange} />
+    },
+  ];
+
 
   return (
     <>
       <div className="p-4">
         <Tabs value={activeTab} className="">
           <TabsHeader
-            className="rounded-none border-blue-gray-50 bg-transparent p-0 w-48"
+            className="rounded-none border-blue-gray-50 bg-transparent p-0 w-48 hidden md:flex"
             indicatorProps={{
               className:
                 "bg-transparent border-b-2 border-gray-900 shadow-none rounded-none",
@@ -216,34 +296,44 @@ export default function Home() {
               <TabPanel key={value} value={value}>
 
                 {/* Flter By */}
-                <div className="py-4 flex items-center justify-between">
-                  <div className="flex items-center gap-4">
+                <div className="py-4 flex flex-col md:flex-row items-center justify-between gap-4">
+
+                  <div className="flex flex-col sm:flex-row items-center gap-4 w-full md:w-auto">
                     <Typography variant="paragraph" color="blue-gray" className="font-medium w-48"
                       {...(undefined as any)}>
                       Filter by:
                     </Typography>
 
-                    <Select size="md" label="Category" className="overflow-hidden"
+                    <Select size="md" label="Category" className="overflow-hidden "
+                    value={payload.taskCategory} 
                       onChange={(value) => {
-                        handleFilterChange("taskStatus", value);
+                        handleFilterChange("taskCategory", value);
                       }}
                       {...(undefined as any)}>
-                      {taskCategories.map((category: any) => (
-                        <Option key={category.id} value={category.id} className="truncate">
-                          {category.name}
+                      {categories.map((category: any) => (
+                        <Option key={category.id} value={category} className="truncate">
+                          {category}
                         </Option>
                       ))}
                     </Select>
 
-                    <Select size="md" label="Due Date" className=" overflow-hidden"      {...(undefined as any)}>
-                      <Option value="asc">Oldest First</Option>
-                      <Option value="desc">Newest First</Option>
-                    </Select>
+                    <DatePicker
+                      selected={startDate}
+                      onChange={handleDateChange}
+                      startDate={startDate}
+                      endDate={endDate}
+                      selectsRange
+                      isClearable
+                      placeholderText="Select date range"
+                      dateFormat="yyyy-MM-dd"
+                      className="border p-2 rounded"
+                    />
                   </div>
 
                   <div className="flex items-center gap-4">
                     <div className="relative">
                       <Input type="text" placeholder="Search" className="w-52 pl-10 rounded-full"
+                       value={payload.taskTitle} 
                         onChange={(e) => {
                           handleFilterChange("taskTitle", e.target.value);
                         }}
@@ -283,6 +373,8 @@ export default function Home() {
         <AddEditTask open={isModalOpen} handleToggle={toggleModal}
           onSave={handleSaveTask} taskEditData={editTaskData} resetEdit={handleResetEditTask} />
       </div>
+
+      <ToastContainer />
     </>
   );
 }
